fix: corrige testes, notebook e model_config para compatibilidade com API refatorada
- test_api.py: fixtures agora patcham _state dict modular em vez de app.state
- test_api.py: usa ChurnMLP em vez de MLP (alias) no mock do modelo
- test_schema.py: muda import para pandera.pandas (elimina FutureWarning)
- test_schema.py: corrige tenure para int em vez de float no RAW_SCHEMA
- model_config.json: adiciona hidden_dims [128,64,32] para API instanciar MLP corretamente
- src/training/train.py: salva hidden_dims no model_config.json ao final do treino
- notebooks/modeling.ipynb: célula et2-save salva hidden_dims no model_config.json

+++ tests/test_schema.py
@@ -2,7 +2,7 @@
 
 import os
 
-import pandera as pa
+import pandera.pandas as pa
 import pytest
 
 from src.data.preprocessing import (
@@ -16,7 +16,7 @@ DATA_PATH = os.getenv("DATA_PATH", "data/raw/Telco_customer_churn.csv")
 
 RAW_SCHEMA = pa.DataFrameSchema(
     {
-        "tenure": pa.Column(float, checks=pa.Check.ge(0), nullable=False),
+        "tenure": pa.Column(int, checks=pa.Check.ge(0), nullable=False),
         "monthly_charges": pa.Column(float, checks=pa.Check.ge(0), nullable=False),
     }
 )
